+++ homes/aarch64-darwin/mariovagomarzal/firefox.nix
@@ -1,6 +1,0 @@
-#######################################
-# Firefox module for mariovagomarzal. #
-#######################################
-{pkgs, ...}: {
-  programs.firefox.package = pkgs.firefox-bin;
-}