+++ lib/default.nix
@@ -4,8 +4,6 @@
 inputs: let
   # Auxiliary function to import libraries.
   importLib = file: import file inputs;
-in
-  # Top-level definitions.
-  (importLib ./core.nix)
-  # Scoped definitions.
-  // {}
+in {
+  modules = importLib ./modules.nix;
+}