+++ modules/home-darwin/firefox.nix
@@ -1,0 +1,6 @@
+##############################
+# Firefox module for Darwin. #
+##############################
+{pkgs, ...}: {
+  programs.firefox.package = pkgs.firefox-bin;
+}