+++ modules/home/ssh.nix
@@ -1,0 +1,21 @@
+##################################
+# SSH configuration home module. #
+##################################
+{...}: {
+  programs.ssh = {
+    enable = true;
+
+    # Configuration for all hosts.
+    addKeysToAgent = "yes";
+    extraConfig = ''
+      AddKeysToAgent yes
+    '';
+
+    matchBlocks = {
+      # GitHub host.
+      "github.com" = {
+        identityFile = "~/.ssh/id_ed25519";
+      };
+    };
+  };
+}